backend/requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
joblib
python-multipart
xgboost
tensorflow
Pillow
scipy
